Rename BuildDacpac to DacpacTool
