+++ MSBuild.Sdk.SqlProj.sln
@@ -7,16 +7,16 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{B0420F9B-A90
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MSBuild.Sdk.SqlProj", "src\MSBuild.Sdk.SqlProj\MSBuild.Sdk.SqlProj.csproj", "{F4A31183-CB24-4D39-B498-8FE4F1A0CCA2}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BuildDacpac", "src\BuildDacpac\BuildDacpac.csproj", "{94A9E682-EF0D-4B51-8589-30F241CF5837}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D25C8812-D92E-41D5-9BC9-30AFCF45063B}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{FD3F6B38-77EB-444F-B9B9-353F26EE8252}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dacpactool", "src\DacpacTool\Dacpactool.csproj", "{78CA944A-928F-48FA-B29F-C5DC96E67684}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildDacpac.Tests", "test\BuildDacpac.Tests\BuildDacpac.Tests.csproj", "{E09D484E-594E-4614-8497-47B01D39087D}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{A3A4D55B-4484-4501-894F-0B07708A9A1D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DacpacTool.Tests", "test\DacpacTool.Tests\DacpacTool.Tests.csproj", "{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -40,38 +40,38 @@ Global
 		{F4A31183-CB24-4D39-B498-8FE4F1A0CCA2}.Release|x64.Build.0 = Release|Any CPU
 		{F4A31183-CB24-4D39-B498-8FE4F1A0CCA2}.Release|x86.ActiveCfg = Release|Any CPU
 		{F4A31183-CB24-4D39-B498-8FE4F1A0CCA2}.Release|x86.Build.0 = Release|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Debug|x64.Build.0 = Debug|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Debug|x86.Build.0 = Debug|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Release|Any CPU.Build.0 = Release|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Release|x64.ActiveCfg = Release|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Release|x64.Build.0 = Release|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Release|x86.ActiveCfg = Release|Any CPU
-		{94A9E682-EF0D-4B51-8589-30F241CF5837}.Release|x86.Build.0 = Release|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Debug|x64.Build.0 = Debug|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Debug|x86.Build.0 = Debug|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Release|x64.ActiveCfg = Release|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Release|x64.Build.0 = Release|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Release|x86.ActiveCfg = Release|Any CPU
-		{E09D484E-594E-4614-8497-47B01D39087D}.Release|x86.Build.0 = Release|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Debug|x64.Build.0 = Debug|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Debug|x86.Build.0 = Debug|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Release|Any CPU.Build.0 = Release|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Release|x64.ActiveCfg = Release|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Release|x64.Build.0 = Release|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Release|x86.ActiveCfg = Release|Any CPU
+		{78CA944A-928F-48FA-B29F-C5DC96E67684}.Release|x86.Build.0 = Release|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Debug|x64.Build.0 = Debug|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Debug|x86.Build.0 = Debug|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Release|x64.ActiveCfg = Release|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Release|x64.Build.0 = Release|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Release|x86.ActiveCfg = Release|Any CPU
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{F4A31183-CB24-4D39-B498-8FE4F1A0CCA2} = {B0420F9B-A902-4E9D-8B29-55B4F626483B}
-		{94A9E682-EF0D-4B51-8589-30F241CF5837} = {B0420F9B-A902-4E9D-8B29-55B4F626483B}
-		{E09D484E-594E-4614-8497-47B01D39087D} = {FD3F6B38-77EB-444F-B9B9-353F26EE8252}
+		{78CA944A-928F-48FA-B29F-C5DC96E67684} = {B0420F9B-A902-4E9D-8B29-55B4F626483B}
+		{2EC1BB4C-C8A1-4C7E-AAE9-24E22EB0A8C4} = {A3A4D55B-4484-4501-894F-0B07708A9A1D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {832DCF12-9633-4AEB-A9D8-444EDF649047}
