+++ src/MSBuild.Sdk.SqlProj/Sdk/Sdk.targets
@@ -90,18 +90,18 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.Managed.targets" Condition="'$(MSBuildAssemblyVersion)' >= '16.0' And '$(ManagedLanguageTargetsGotImported)' != 'true'" />
 
   <!--
-    Validates that the environment is setup correctly for the BuildDacpac tool to run. Currently we target .NET Core 2.1 and 3.1 and we pick the highest version based on
+    Validates that the environment is setup correctly for the DacpacTool to run. Currently we target .NET Core 2.1 and 3.1 and we pick the highest version based on
     the BundledNETCoreAppTargetFrameworkVersion property. This property is set by the SDK, so we'll need to ensure that the SDK that is being used is compatible with one
     of the supported target frameworks.
   -->
   <Target Name="ValidateEnvironment">
     <PropertyGroup>
-      <BuildDacpacTfm>netcoreapp$(BundledNETCoreAppTargetFrameworkVersion)</BuildDacpacTfm>
-      <BuildDacpacExe>$(MSBuildThisFileDirectory)../tools/$(BuildDacpacTfm)/BuildDacpac.dll</BuildDacpacExe>
+      <DacpacToolTfm>netcoreapp$(BundledNETCoreAppTargetFrameworkVersion)</DacpacToolTfm>
+      <DacpacToolExe>$(MSBuildThisFileDirectory)../tools/$(DacpacToolTfm)/DacpacTool.dll</DacpacToolExe>
     </PropertyGroup>
-    <Error Condition="!Exists($(BuildDacpacExe))"
+    <Error Condition="!Exists($(DacpacToolExe))"
            Text="The current SDK bundles target framework 'netcoreapp$(BundledNETCoreAppTargetFrameworkVersion)' which is not supported by MSBuild.Sdk.SqlProj. Either install a .NET Core SDK that supports 2.1 or 3.1, or file an issue at https://github.com/jmezach/MSBuild.Sdk.SqlProj/ to add support." />
-    <Message Importance="normal" Text="Using target framework $(BuildDacpacTfm) to run BuildDacpac" />
+    <Message Importance="normal" Text="Using target framework $(DacpacToolTfm) to run DacpacTool" />
   </Target>
 
   <!--
@@ -145,12 +145,11 @@
       <SqlCmdVariableArguments>@(SqlCmdVariable->'-sc %(Identity)', ' ')</SqlCmdVariableArguments>
       <PreDeploymentScriptArgument>@(PreDeploy->'--predeploy %(Identity)', ' ')</PreDeploymentScriptArgument>
       <PostDeploymentScriptArgument>@(PostDeploy->'--postdeploy %(Identity)', ' ')</PostDeploymentScriptArgument>
-      <BuildDacpacTfm>netcoreapp$(BundledNETCoreAppTargetFrameworkVersion)</BuildDacpacTfm>
-      <BuildDacpacCommand>dotnet $(MSBuildThisFileDirectory)../tools/$(BuildDacpacTfm)/BuildDacpac.dll build $(OutputPathArgument) $(MetadataArguments) $(SqlServerVersionArgument) $(InputFileArguments) $(ReferenceArguments) $(SqlCmdVariableArguments) $(PropertyArguments) $(PreDeploymentScriptArgument) $(PostDeploymentScriptArgument)</BuildDacpacCommand>
+      <DacpacToolCommand>dotnet $(DacpacToolExe) build $(OutputPathArgument) $(MetadataArguments) $(SqlServerVersionArgument) $(InputFileArguments) $(ReferenceArguments) $(SqlCmdVariableArguments) $(PropertyArguments) $(PreDeploymentScriptArgument) $(PostDeploymentScriptArgument)</DacpacToolCommand>
     </PropertyGroup>
     <!-- Run it, except during design-time builds -->
-    <Message Importance="Low" Text="Running command: $(BuildDacpacCommand)" />
-    <Exec Command="$(BuildDacpacCommand)"
+    <Message Importance="Low" Text="Running command: $(DacpacToolCommand)" />
+    <Exec Command="$(DacpacToolCommand)"
           Condition="'$(DesignTimeBuild)' != 'true' AND '$(BuildingProject)' == 'true'"/>
 
     <!-- Copy any referenced .dacpac packages to the output folder -->
@@ -159,8 +158,7 @@
 
   <Target Name="Publish" DependsOnTargets="Build">
     <PropertyGroup>
-      <DacpacToolTfm>netcoreapp$(BundledNETCoreAppTargetFrameworkVersion)</DacpacToolTfm>
-      <DacpacToolCommand>dotnet $(MSBuildThisFileDirectory)../tools/$(BuildDacpacTfm)/BuildDacpac.dll deploy</DacpacToolCommand>
+      <DacpacToolCommand>dotnet $(DacpacToolExe) deploy</DacpacToolCommand>
     </PropertyGroup>
     <Message Importance="Low" Text="Running command: $(DacpacToolCommand)" />
     <Exec Command="$(DacpacToolCommand)" />
