+++ test/DacpacTool.Tests/DacpacHeaderParser/MetaData.cs
@@ -1,4 +1,4 @@
-﻿namespace MSBuild.Sdk.SqlProj.BuildDacpac.Tests.DacpacHeaderParser
+﻿namespace MSBuild.Sdk.SqlProj.DacpacTool.Tests.DacpacHeaderParser
 {
     public class Metadata
     {
