+++ test/DacpacTool.Tests/PackageBuilderTests.cs
@@ -8,7 +8,7 @@ using Microsoft.SqlServer.Dac.Model;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Shouldly;
 
-namespace MSBuild.Sdk.SqlProj.BuildDacpac.Tests
+namespace MSBuild.Sdk.SqlProj.DacpacTool.Tests
 {
     [TestClass]
     public class PackageBuilderTest
